+++ src/foundation/components/vehicles/aerial-sports/ParagliderVoxel.ts
@@ -1,5 +1,5 @@
 import * as THREE from 'three';
-import LegacyPilotVoxel, { PilotVoxelOptions } from '../../characters/LegacyPilotVoxel';
+import { PilotVoxel, PilotVoxelOptions } from '../../characters/PilotVoxel';
 import Glider, { GliderOptions } from '../components/Glider';
 import GuiHelper from '../../../utils/gui';
 import { IVehicleWithGUI } from '../../../types/IVehicle';
@@ -19,13 +19,13 @@ export interface ParagliderVoxelOptions {
  * Implements IVehicleWithGUI for consistent vehicle interface across the application.
  *
  * Note: This class doesn't extend a base component class because it's a composite
- * that assembles two different component types (Glider and LegacyPilotVoxel) with
+ * that assembles two different component types (Glider and PilotVoxel) with
  * different APIs. Instead, it implements the IVehicle interface for consistency.
  */
 export class ParagliderVoxel implements IVehicleWithGUI {
   private mesh?: THREE.Object3D;
   private glider!: Glider;
-  private pilot!: LegacyPilotVoxel;
+  private pilot!: PilotVoxel;
   private pilotMesh!: THREE.Object3D;
   private options: ParagliderVoxelOptions;
 
@@ -48,7 +48,7 @@ export class ParagliderVoxel implements IVehicleWithGUI {
     this.mesh.add(wing);
 
     // Create voxel pilot
-    this.pilot = new LegacyPilotVoxel(this.options.pilot);
+    this.pilot = new PilotVoxel(this.options.pilot);
     this.pilotMesh = await this.pilot.load();
     this.pilotMesh.position.x = 350;
     this.pilotMesh.position.y = -600;